OOP
Learned a little more about the utility of OOP. Created a new Cart class to abstract some of the information in the req.session.cart. I did this to finally create subtotals and totals that include user sales tax rates.  Using classes to organize information works so well and It seems so sophisticated, I'm proud of myself when I do it.